refactor: wallets query options 공통화
- wallet.queries에서 walletKeys 분리
