+++ src/features/home/api/wallet.queries.ts
@@ -5,25 +5,11 @@ import {
   useInfiniteQuery,
 } from '@tanstack/react-query'
 import { getWallets, deleteWalletCard } from './wallet.api'
+import { walletKeys, walletsInfiniteQueryOptions } from './wallet.query-options'
 import type { GetWalletsParams } from '../model/wallet.types'
 
-export const walletKeys = {
-  all: ['wallets'] as const,
-  lists: () => [...walletKeys.all, 'list'] as const,
-  list: (params: GetWalletsParams = {}) =>
-    [...walletKeys.lists(), params] as const,
-}
-
 export function useWallets(params: Omit<GetWalletsParams, 'cursorId'> = {}) {
-  return useInfiniteQuery({
-    queryKey: walletKeys.list(params),
-    queryFn: ({ pageParam }) =>
-      getWallets({ ...params, cursorId: pageParam as number | undefined }),
-    initialPageParam: undefined as number | undefined,
-    getNextPageParam: (lastPage) =>
-      lastPage.pagination?.has_next ? lastPage.pagination.cursorId : undefined,
-    refetchOnWindowFocus: false,
-  })
+  return useInfiniteQuery(walletsInfiniteQueryOptions(params, getWallets))
 }
 
 export function useWalletList(params: GetWalletsParams = {}) {
