+++ src/features/home/ui/index.ts
@@ -7,11 +7,12 @@ export {
   type BusinessCardListProps,
 } from './business-card-list'
 export { SearchInput, type SearchInputProps } from './search-input'
-export { MyCardButton, type MyCardButtonProps } from './my-card-button'
 export { FAB, type FABProps } from './fab'
-export { FABMenu, type FABMenuProps, type FABMenuItem } from './fab-menu'
+export { FABMenu, type FABMenuProps } from './fab-menu'
 export { OcrModeDialog, type OcrModeDialogProps } from './ocr-mode-dialog'
 export {
   EmptyCardPlaceholder,
   type EmptyCardPlaceholderProps,
 } from './empty-card-placeholder'
+export { HomeHeader } from './home-header'
+export { HomeContent, type HomeContentProps } from './home-content'
